+++ packages/connect/src/index.ts
@@ -3,6 +3,7 @@ import type { PeerId } from '@libp2p/interface-peer-id'
 import type { Initializable, Components } from '@libp2p/interfaces/components'
 import type { Startable } from '@libp2p/interfaces/startable'
 import type { Connection, MultiaddrConnection } from '@libp2p/interface-connection'
+import { CustomEvent } from '@libp2p/interfaces/events'
 
 import errCode from 'err-code'
 import Debug from 'debug'
@@ -32,6 +33,7 @@ import { EntryNodes } from './base/entry.js'
 import { WebRTCUpgrader } from './webrtc/upgrader.js'
 import { UpnpManager } from './base/upnp.js'
 import { timeout } from '@hoprnet/hopr-utils'
+import { cleanExistingConnections } from './utils/index.js'
 
 const DEBUG_PREFIX = 'hopr-connect'
 const log = Debug(DEBUG_PREFIX)
@@ -237,13 +239,28 @@ class HoprConnect implements Transport, Initializable, Startable {
   private async dialWithRelay(relay: PeerId, destination: PeerId, options: DialOptions): Promise<Connection> {
     log(`Dialing ${chalk.yellow(`/p2p/${relay.toString()}/p2p-circuit/p2p/${destination.toString()}`)}`)
 
-    let maConn = await this.getConnectComponents().getRelay().connect(relay, destination, options)
+    let conn: Connection | undefined
+
+    let maConn = await this.getConnectComponents()
+      .getRelay()
+      .connect(
+        relay,
+        destination,
+        () => {
+          if (conn) {
+            ;(this.components as Components).getUpgrader().dispatchEvent(
+              new CustomEvent(`connectionEnd`, {
+                detail: conn
+              })
+            )
+          }
+        },
+        options
+      )
 
     if (maConn == undefined) {
       throw Error(`Could not establish relayed connection.`)
     }
-
-    let conn: Connection
 
     try {
       conn = await options.upgrader.upgradeOutbound(maConn)
@@ -254,6 +271,9 @@ class HoprConnect implements Transport, Initializable, Startable {
       // want to log it for debugging purposes
       throw err
     }
+
+    // Not supposed to throw any exception
+    cleanExistingConnections(this.components as Components, conn.remotePeer, conn.id, error)
 
     // Merges all tags from `maConn` into `conn` and then make both objects
     // use the *same* array
@@ -282,13 +302,29 @@ class HoprConnect implements Transport, Initializable, Startable {
   ): Promise<Connection> {
     log(`Dialing ${chalk.yellow(ma.toString())}`)
 
-    const maConn = await TCPConnection.create(ma, options)
+    let conn: Connection | undefined
+    const maConn = await TCPConnection.create(
+      ma,
+      () => {
+        if (conn) {
+          ;(this.components as Components).getUpgrader().dispatchEvent(
+            new CustomEvent(`connectionEnd`, {
+              detail: conn
+            })
+          )
+        }
+      },
+      options
+    )
 
     verbose(
       `Establishing a direct connection to ${maConn.remoteAddr.toString()} was successful. Continuing with the handshake.`
     )
 
-    const conn = await timeout(DEFAULT_CONNECTION_UPGRADE_TIMEOUT, () => options.upgrader.upgradeOutbound(maConn))
+    conn = await timeout(DEFAULT_CONNECTION_UPGRADE_TIMEOUT, () => options.upgrader.upgradeOutbound(maConn))
+
+    // Not supposed to throw any exception
+    cleanExistingConnections(this.components as Components, conn.remotePeer, conn.id, error)
 
     // Assign various connection properties once we're sure that public keys match,
     // i.e. dialed node == desired destination