+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 - Allow `info` command before node startup has finished ([#4273](https://github.com/hoprnet/hoprnet/pull/4273))
 - Turn libp2p dual DHT into single DHT by forking DHT package in order to avoid a memory leak https://github.com/hoprnet/hoprnet/pull/4288
 - Turn db operations into `zero-copy` operations ([#4293](https://github.com/hoprnet/hoprnet/pull/4293))
+- Close existing connections once there is new one ([#4281](https://github.com/hoprnet/hoprnet/pull/4281))
+- Properly remove closed connections from libp2p's `ConnectionManager`([#4281](https://github.com/hoprnet/hoprnet/pull/4281))
 
 ---